win, x: move x_setCommonPropertiesForAnyWindow() to win
win.c handles exacly common things, where customization is done
under struct WmOps, so the function more suits there.

If there is a need to move X related stuff to a separate module,
then it makes sense to look at that independently of WM_NO and
WM_TWM code and move it to more self-descriptive file.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>