gui: skip extra winlist initializations
The winlist is initialized inside of win's initWinlist, so move it
totally there. There is only one call of that function from gui.c,
so from the codeflow it is basically the same place.

But the patch changes the behaviour a bit:

1) in case if winlist was initialized before, it does not blindly
assign NULL, but frees the memory.
2) if WM specific code fails in the middle, it frees the data before
debug output and returning, so in the error case the output will
differ.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>