ewmh: fold window property calls
Small refactoring. There is a common pattern in several places:

- check one property;
  if it does not exist
     check another property.

Make a util helper for that, which gets 2 properties as arguments
and returns the first possible.

Second level is fold ewmh_getCurrentDesktop() and
ewmh_getDesktopOfWindow() since their code is almost the same but
checks different properties.

NOTE: in the code e->wmname can get or dynamically allocated
property string, or static string. For the dynamic case I see here a
leak.

Signed-off-by: Yauheni Kaliuta <y.kaliuta@gmail.com>