make sure all test X clients connect to the Xvfb instance
Export the DISPLAY variable immediately after starting Xvfb so that
any X clients started after that, including the recently added xdotool,
will be sure to use it. Without this change, setting DISPLAY (without
exporting it) in alttab_simple_cycle() does not work when that variable
is not already exported, e.g. when running in an isolated environment
like a container or a chroot environment.